backend/requirements.txt
africastalking==1.2.9
asgiref==3.9.0
async-timeout==5.0.1
certifi==2025.6.15
charset-normalizer==3.4.2
channels==4.2.2
channels_redis==4.2.1
Django==5.2.4
django-allauth==65.9.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
idna==3.10
msgpack==1.1.1
Pillow-11.3.0
PyJWT==2.9.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
redis==6.2.0
requests==2.32.4
schema==0.7.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchman==1.3.0
openrouteservice
twilio
